When the user clicks, sort by descending first

+++ src/Request.kt
@@ -29,7 +29,7 @@ open class Request(val template: String, val words: List<String?>, val learnBori
     var endpointOverride: String? = null
     var montoyaReq: HttpRequestResponse? = null
     var autoFixContentLength: Boolean = true
-    var anomalyRank: Float? = null
+    var anomalyRank: Int? = null
 
     private val attributes: HashMap<String, Any> = HashMap()
